refactor: add NotImplementedError for driver functions that do nothing (#5413)
Many of the driver functions do not support or implement certain
functions, yet when executed they return successfully, and may give
the false impression that they work.

+++ test/functional/schema-builder/custom-db-and-schema-sync.ts
@@ -97,7 +97,9 @@ describe("schema builder > custom-db-and-schema-sync", () => {
         await queryRunner.dropTable(albumMetadata.tablePath, true, false);
 
         // drop created database
-        await queryRunner.dropDatabase("secondDB", true);
+        if (!(connection.driver instanceof PostgresDriver)) {
+            await queryRunner.dropDatabase("secondDB", true);
+        }
 
         await queryRunner.release();
